Fix test case id set
