+++ src/utils/utils.ts
@@ -5,6 +5,7 @@ import remarkRehype from "remark-rehype";
 import rehypeStringify from "rehype-stringify";
 import rehypeSanitize from "rehype-sanitize";
 import supersub from "remark-supersub";
+import { LanguageName } from "@uiw/codemirror-extensions-langs";
 
 export const makeMap = <T, K extends keyof T & string>(l: T[], k: K) =>
   new Map(l.map((q) => [q[k], q]));
@@ -33,6 +34,26 @@ export const pasteText = (
     "end",
   );
 };
+
+export function getLanguage(language: string): LanguageName | undefined {
+  const l = language.toLowerCase();
+
+  if (!l.includes("c++") && (l.includes("gcc") || l.includes("clang"))) {
+    return "c";
+  }
+  if (l.includes("c++")) {
+    return "cpp";
+  }
+  if (l.includes("java")) {
+    return "java";
+  }
+
+  if (l.includes("python")) {
+    return "python";
+  }
+  // default to markdown
+  return undefined;
+}
 
 export const bracket = (
   inputRef: HTMLTextAreaElement | null | undefined,
